+++ Src/Public/Invoke-AsBuiltReport.Microsoft.AD.ps1
@@ -22,7 +22,6 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     )
 
     #Requires -RunAsAdministrator
-    #Requires -PSEdition Core
 
     if ($psISE) {
         Write-Error -Message $reportTranslate.InvokeAsBuiltReportMicrosoftAD.PwshISE
@@ -36,7 +35,7 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     Write-Host $reportTranslate.InvokeAsBuiltReportMicrosoftAD.ReportModuleInfo6
 
     # Check the version of the dependency modules
-    $ModuleArray = @('AsBuiltReport.Core', 'Diagrammer.Core', 'PSPKI')
+    $ModuleArray = @('AsBuiltReport.Core', 'AsBuiltReport.Chart', 'Diagrammer.Core', 'PSPKI')
 
     foreach ($Module in $ModuleArray) {
         try {